Refactoring of the request/response architecture
Ferox is going to allow for inbound and outbound request handlers
to be attached to a given route (e.g., /users/:id).

Outbound handlers will receive each message sent by the inbound handlers
and be able to do additional processing on those messages.

One use case for this is webfinger and remotestorage.

Way may get some webfinger request, pull the resource, and send the resource
but remoteStorage needs to add some additonal links to that webfinger resource.

remoteStorage will be able to attach an outbound route handler to the webfinger
route (/.well-known/webfinger?resource=) and add information to the
resource being sent by the webfinger inbound route handler.